agent/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
supabase
twilio
langchain
langchain-openai
langchain-anthropic
langgraph
